Clean up security auth imports
