+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -196,6 +196,13 @@
 - **Summary**: Refactored from 680 to 200 lines by extracting core, collector, analyzer, and config modules. Main orchestrator now imports these modules while preserving functionality.
 
 
+### MODERATE-025: Security Manager Modularization ✅
+- **File**: `src/core/security/security_manager.py`
+- **Status**: Completed
+- **Assigned To**: Agent-2
+- **Completion Date**: 2025-08-24
+- **Summary**: Refactored monolithic security manager into core, auth, and config modules. Orchestrator now imports these modules and original file removed.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
