+++ src/core/security/security_manager.py
@@ -1,0 +1,11 @@
+from .security_auth import SecurityLevel, Permission, SecurityPolicy, AccessLog
+from .security_config import get_default_isolation_rules
+from .security_manager_core import WorkspaceSecurityManager
+
+__all__ = [
+    "SecurityLevel",
+    "Permission",
+    "SecurityPolicy",
+    "AccessLog",
+    "WorkspaceSecurityManager",
+]
