+++ src/core/workspace_manager.py
@@ -8,7 +8,7 @@ from .workspace_config import (
     WorkspaceConfigManager,
 )
 from .workspace_creator import WorkspaceStructureManager
-from .workspace_validator import (
+from .security.security_manager import (
     SecurityLevel,
     Permission,
     SecurityPolicy,
